+++ desc/geometry/core.py
@@ -284,7 +284,7 @@ class Curve(IOAble, Optimizable, ABC):
             coords, knots=knots, method=method, name=name, basis="xyz"
         )
 
-    def to_FourierRZ(self, N=None, grid=None, NFP=None, name=""):
+    def to_FourierRZ(self, N=None, grid=None, NFP=None, sym=False, name=""):
         """Convert Curve to FourierRZCurve representation.
 
         Note that some types of curves may not be representable in this basis.
@@ -299,6 +299,9 @@ class Curve(IOAble, Optimizable, ABC):
         NFP : int
             Number of field periods, the curve will have a discrete toroidal symmetry
             according to NFP.
+        sym : bool, optional
+            whether the curve is stellarator-symmetric or not. Default
+            is False.
         name : str
             name for this curve
 
@@ -314,7 +317,41 @@ class Curve(IOAble, Optimizable, ABC):
         if grid is None:
             grid = LinearGrid(N=2 * N + 1, NFP=NFP)
         coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return FourierRZCurve.from_values(coords, N=N, NFP=NFP, basis="xyz", name=name)
+        return FourierRZCurve.from_values(
+            coords, N=N, NFP=NFP, basis="xyz", name=name, sym=sym
+        )
+
+    def to_FourierPlanar(self, N=None, grid=None, name=""):
+        """Convert Curve to FourierPlanarCurve representation.
+
+        Note that some types of curves may not be representable in this basis.
+        In this case, a least-squares fit will be done to find the
+        planar curve that best represents the curve.
+
+        Parameters
+        ----------
+        N : int
+            Fourier resolution of the new FourierPlanarCurve representation.
+        grid : Grid, int or None
+            Grid used to evaluate curve coordinates on to fit with FourierPlanarCurve.
+            If an integer, uses that many equally spaced points.
+        name : str
+            name for this curve
+
+        Returns
+        -------
+        curve : FourierPlanarCurve
+            New representation of the curve parameterized by Fourier series for
+            minor radius r in a plane specified by a center position and normal
+            vector.
+
+        """
+        from .curve import FourierPlanarCurve
+
+        if grid is None:
+            grid = LinearGrid(N=2 * N + 1)
+        coords = self.compute("x", grid=grid, basis="xyz")["x"]
+        return FourierPlanarCurve.from_values(coords, N=N, basis="xyz", name=name)
 
 
 class Surface(IOAble, Optimizable, ABC):